Convert add_to_log() to the new logging architecture
This is slightly breaking the backwards compatibility. From now on,
opening the 'Manage stamps' is no longer logged as a separate 'manage
stamps' action (when legavy logging is on). Instead, the standard course
module viewed event is triggered with the property other[viewmode] set
to 'manage'. So it is stored as 'view' action in the legacy log. This
better describes the essence of the event.