Trigger event when a user receives a stamp
This seems to be more useful that the current 'stamp added' event. The
main reason for introducing this one is to be able to use the Level Up!
block together with the Stamp collection. For the reporting and
analysing purposes, both events are equal and either of them can be
used.